web: completely redo act layouting
The former hierarchy-based layout was replaced with a part-based one,
mainly to allow for easy diffing between two documents: it's easier to diff
two (mostly) linear vectors than simultaneously walking two trees.

This also allows us to create unit tests for the act tree walking even if
we want to radically change the actual HTML layout and CSS stuff in the
future.

The code is much cleaner now that linearization and actual HTML stuff is
separate. The two new rendering contexts are also cleaner than the single
one was before.

In the future, we may even go with some streaming-based solution, because
all we do with the 'output' variable is push at the end of it.

Other changes:
- RenderElementContext methods now take an owned self. Cloning before
  calling these methods is better practice than what we had earlier,
  especially when chained.