web/diff: add anchor strings to right side too
This will have a ton of elements with the same ID, but at least links will
work in cases where only the right side has the element (common in case of
insertions).

A better solution would be to add the ID only when the element exists on
the right side only.